Paginação

+++ src/main/java/br/com/exercicio4/domain/service/UserService.java
@@ -1,7 +1,13 @@
 package br.com.exercicio4.domain.service;
 
+import br.com.exercicio4.domain.ValidationException;
 import br.com.exercicio4.domain.model.User;
 import br.com.exercicio4.infra.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,26 +21,45 @@ public class UserService {
         this.repository = repository;
     }
 
+    @Transactional
     public User create(User user) {
+        if (repository.existsByEmail(user.getEmail())) {
+            throw new ValidationException("Já existe um usuário com esse e-mail.");
+        }
         return repository.save(user);
     }
 
-    public List<User> listAll() {
-        return repository.findAll();
+    public Page<User> listAll(Pageable pageable) {
+        return repository.findAll(pageable);
     }
 
     public User findById(Long id) {
         return repository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Usuário não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado com ID: " + id));
     }
 
+    @Transactional
     public User update(Long id, String name, String email) {
         User user = findById(id);
-        user.update(name, email);
+
+        if (email != null && !email.isBlank() && !email.equals(user.getEmail()) && repository.existsByEmail(email)) {
+            throw new ValidationException("Email já está em uso por outro usuário.");
+        }
+
+        user.updateInfo(name, email);
         return repository.save(user);
     }
 
+    @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Usuário não encontrado com ID: " + id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw e; // será tratado pelo ErrorHandler
+        }
     }
 }
